Rebase on Alpine; don't include host compiler.
Because all of the cross-compilation libraries are included, the host
userspace doesn't really matter. Alpine is based on a BusyBox userspace,
which necessitates POSIX compatibility, but we were almost there. We can
also save a lot of space (and possibly avoid conclusion) if we don't
include a compiler targeting the host.